验证问题修正

+++ GeneralKit/Attributes/RuleAttribute.cs
@@ -35,19 +35,19 @@ namespace GeneralKit
         /// 大于
         /// <code/>用于数值类型
         /// </summary>
-        internal decimal? greater;
+        internal double? greater;
 
         /// <summary>
         /// 小于
         /// <code/>用于数值类型
         /// </summary>
-        internal decimal? less;
+        internal double? less;
 
         /// <summary>
         /// 小于
         /// <code/>用于数值类型
         /// </summary>
-        internal decimal? equal;
+        internal double? equal;
 
         /// <summary>
         /// 自定义正则表达式
@@ -63,9 +63,9 @@ namespace GeneralKit
         public int MinLength { get => minLength ?? 0; set => minLength = value; }
         public string Error { get => error; set => error = value; }
         public bool AllowEmpty { get => allowEmpty ?? true; set => allowEmpty = value; }
-        public decimal Greater { get => greater ?? 0; set => greater = value; }
-        public decimal Less { get => less ?? 0; set => less = value; }
-        public decimal Equal { get => equal ?? 0; set => equal = value; }
+        public double Greater { get => greater ?? 0; set => greater = value; }
+        public double Less { get => less ?? 0; set => less = value; }
+        public double Equal { get => equal ?? 0; set => equal = value; }
         public string RegExp { get => regExp; set => regExp = value; }
         public ExpType ExpType { get => expType ?? ExpType.None; set => expType = value; }
     }
